% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_labs.R
\name{check.labs}
\alias{check.labs}
\title{Check variable and value labels}
\usage{
check.labs(vari)
}
\arguments{
\item{vari}{Variable to be checked}
}
\value{
list with four elements:
\itemize{
\item var_lab - variable label (NA if not defined)
\item val_lab - data frame with values and value labels, NA if not defined
\item val_lab_ok - logical. FALSE if there are more values than value labels, otherwise TRUE. NA if value labels are not defined.
}
}
\description{
Helper function to determine if there is a defined variable label and value labels. Also checks if there are more unique values than defined value labels.
}
