% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.R
\name{chisq.flex}
\alias{chisq.flex}
\title{Claculate chi-square test with post-hocs}
\usage{
chisq.flex(data, form, post_hoc_chi = TRUE)
}
\arguments{
\item{data}{labelled data frame}

\item{form}{formula with two variables}

\item{post_hoc_chi}{logical. If TRUE (default), post-hoc chi-square tests are calculated}
}
\value{
Flextable with results
}
\description{
Claculate chi-square test with post-hocs
}
