% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{cor.flex}
\alias{cor.flex}
\title{Correlation matrix from data.frame}
\usage{
cor.flex(
  df,
  na = "listwise",
  type = "pearson",
  labs = "numbered",
  diag = "bottom",
  lang = "hr"
)
}
\arguments{
\item{df}{data.frame}

\item{na}{how will missing values be handled ("listwise" (default), "pairwise")}

\item{type}{type of correlation ("pearson" (default), "spearman")}

\item{labs}{type of table:
\itemize{
\item "numbered" - numbered variable labels an numbers as column headings
\item "asis" - variable labels in rows and column headings
\item "names" - variable names in rows and column headings
}}

\item{diag}{table looks ("bottom" (default), "top", "all")}

\item{lang}{language ("hr" (default), "en")}
}
\value{
list with two elements:
\itemize{
\item $matrix - flextable object
\item $N - number of participants (range if na="pairwise")
}
}
\description{
Creates correlation matrix from data.frame based on Hmisc::rcorr
}
