% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_descriptives2.R
\name{des.flex}
\alias{des.flex}
\title{Provide basic descriptives}
\usage{
des.flex(
  data,
  vars,
  param_set = "standard",
  by = NA,
  by_total = TRUE,
  deci = 1,
  option = r.flex.opts,
  title = ""
)
}
\arguments{
\item{data}{data.frame with variables}

\item{vars}{variables as formula. e.g. 1 ~ var1 + var2}

\item{param_set}{list of parameters with order in the table (as strings):
\itemize{
\item default - "standard"
\item "simple" - N, Min, Max, M, SD
\item "standard" - N, M, SD, Min, Q1, C, Q3, Max, Skew, Kurt, SW, SWp
\item "report" - N, M, SD, Min, Q1, C, Q3, Max
\item other available statistics - missing values(Miss), Kolmogorov-Smirnov (KS, KSp), standard errors for Skewness and Kurtosis (SEskew, SEkurt)
}}

\item{by}{section variable, default=FALSE, otherwise put variable name in quotes, e.g. "section_var". Section variable should have labels}

\item{by_total}{boolean. If there is a section variable, should the total result be included in descriptives (default=TRUE)}

\item{deci}{number of decimals for M, everything else is done automatically (default=1)}

\item{option}{options variable. To change, modify r.flex.opts variable created by im_on_fire()}

\item{title}{table title. If not set it will be determined automatically}
}
\value{
list with elements
\itemize{
\item type - table type - used for inserting in word document
\item title - used for table title. Can be set manually or automatically
\item section - section variable title
\item table - flextable with results
\item tab.df - results as data.frame
\item orientation - suggested page orientation (P/L)
}
}
\description{
Simple descriptive statistics
}
