% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_labs.R
\name{get.var.name}
\alias{get.var.name}
\title{Get variable name or variable label}
\usage{
get.var.name(data, vari)
}
\arguments{
\item{data}{data.frame}

\item{vari}{variable as string}
}
\value{
variable name or variable label if existing as string
}
\description{
helper function that returns variable name or variable label if existing as string. Used for values in tables.
}
