% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRA_flex.R
\name{hra.flex}
\alias{hra.flex}
\title{Hierarhical regression analysis}
\usage{
hra.flex(data, ..., option = r.flex.opts)
}
\arguments{
\item{data}{dataset with variables}

\item{...}{one or more formulas separated by comma (e.g out ~ v1 + v2, out ~ v3 + v4). Each model should not contain variables from previous step}

\item{option}{r.flex.opts}
}
\value{
list with two elements: a flextable and data frame with results
}
\description{
Hierarhical regression analysis
}
