% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_on_fire.R
\name{im_on_fire}
\alias{im_on_fire}
\title{Set up everything needed to start an analysis}
\usage{
im_on_fire()
}
\value{
environment setup and variables:
\itemize{
\item working directory
\item directories in working directory: Data, Pictures, Templates, Outputs
\item downloaded file template.docx into Templates directory
\item global variable in r.flex.opts environment \strong{doc} based on template.docx to be used with officer
\item \strong{r.flex.opts} environment which is used in functions for options:
\itemize{
\item lang - language - defaults to "hr", can be changed to "en"
\item d.p - decimal point type - defaults to ",", can be changed to "."
\item lead.zero (boolean) - will the leading zero be shown in numbers. e.g. 0.05 or .05. Defaults to TRUE
\item p.type - how do you show p values. Options are
\itemize{
\item "<>" - default - they will be shown as compared to critical points (<0.05, <0.01, <0.001)
\item "exact" - shown as exact values
\item "star" - shown as stars as compared to critical points (<0.05, 0.01, 0.001)
}
\item tab.caption (boolean) - should table captions be put in word file. Defaults to TRUE
}
}
}
\description{
This function will do necessary steps to set up an environment for an analysis
Basically, it sets up a working directory, creates options variable in global environment, creates folders and downloads template file for officer package.
}
\details{
You should really start with this
}
