% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{nfac.flex}
\alias{nfac.flex}
\title{Determine how many components to extract in PCA}
\usage{
nfac.flex(data)
}
\arguments{
\item{data}{dataset - all variables are analyzed}
}
\value{
list with dataframe and flextable
}
\description{
This is a helper function for pca.flex function and it can be used to determine the number of components manually. It allows the determination of the number of components using various criteria from functions psych::VSS and nFactors::nScree functions
}
