% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_descriptives2.R
\name{r.ks}
\alias{r.ks}
\title{Calculates Kolmogorov-Smirnov test with Lilliefors correction}
\usage{
r.ks(target_variable)
}
\arguments{
\item{target_variable}{numeric variable}
}
\value{
list with statistic and p value
}
\description{
uses nortest package
}
