% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_descriptives2.R
\name{r.kurtosis}
\alias{r.kurtosis}
\title{Calculates excess kurtosis of the variable}
\usage{
r.kurtosis(target_variable)
}
\arguments{
\item{target_variable}{numeric variable}
}
\value{
excess kurtosis
}
\description{
uses SPSS Book of algorithms v20
}
