% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_descriptives2.R
\name{r.skewness}
\alias{r.skewness}
\title{Calculates skewness of the variable}
\usage{
r.skewness(target_variable)
}
\arguments{
\item{target_variable}{numeric variable}
}
\value{
variable skewness
}
\description{
uses Kirk (2008) method for adjusted Fisher-Pearson standardized moment coefficient.
}
\details{
This is the same method Excel uses when calculating skewness.
}
