% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_descriptives2.R
\name{r.skewness.se}
\alias{r.skewness.se}
\title{Calculates standard error for the skewness of the variable}
\usage{
r.skewness.se(target_variable)
}
\arguments{
\item{target_variable}{numeric variable}
}
\value{
standard error for variable skewness
}
\description{
uses SPSS book of algorithms
}
