% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_descriptives2.R
\name{r.sw}
\alias{r.sw}
\title{Calculates Shapiro-Wilk test}
\usage{
r.sw(target_variable)
}
\arguments{
\item{target_variable}{numeric variable}
}
\value{
list with statistic and p value
}
\description{
The maximum data length is 5000 points by definition. If the sample size is larger than 5000 (without NAs), function selects a random sample with 5000 cases and issues a warning.
}
