% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflexR-package.R
\docType{package}
\name{reflexR-package}
\alias{reflexR}
\alias{reflexR-package}
\title{reflexR: FLEXTABLE BASED STATISTICS}
\description{
A flextable based package for doing statistical reports.
}
\author{
\strong{Maintainer}: Miroslav Rajter \email{miroslav.rajter@gmail.com} (\href{https://orcid.org/0000-0002-2608-3196}{ORCID})

}
\keyword{internal}
