% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_on_fire.R
\name{set.r.flex.opts}
\alias{set.r.flex.opts}
\title{Set r.flex.opts variable}
\usage{
set.r.flex.opts(
  type = "fixed",
  lang = NULL,
  d.p = NULL,
  lead.zero = NULL,
  p.type = NULL,
  tab.caption = NULL
)
}
\arguments{
\item{type}{if "interactive" it will ask you for each parameter. Otherwise (default), the parameters need to be defined as function parameters.}

\item{lang}{default language for the package. Currently "hr" or "en"}

\item{d.p}{set decimal point to "." or ","}

\item{lead.zero}{would you like to have a leading zero (T/F)}

\item{p.type}{How should p values be displayed("<>", "exact", "star")}

\item{tab.caption}{Should table captions be put in word file (T/F)}
}
\value{
r.flex.opts environment used in various functions
}
\description{
Used as a direct way to set the options variable used in other functions. The preferred way is by using im_on_fire() function.
This should be used when you want to work without other perks of im_on_fire().
}
