% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{small.t.flex}
\alias{small.t.flex}
\title{single t-test. A helper function for ttest.flex}
\usage{
small.t.flex(
  data,
  form,
  multi,
  groups,
  effect = "d",
  type = "two.sided",
  deci = 2,
  conf.level = 0.95
)
}
\arguments{
\item{data}{a parameter}

\item{form}{a parameter}

\item{multi}{a parameter}

\item{groups}{a parameter}

\item{effect}{a parameter}

\item{type}{a parameter}

\item{deci}{a parameter}

\item{conf.level}{a parameter}
}
\value{
single dataset for t-test
}
\description{
single t-test. A helper function for ttest.flex
}
