% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals.R
\name{star.pval.apa}
\alias{star.pval.apa}
\title{Converts number to APA 7 format and adds optional stars for pvalues}
\usage{
star.pval.apa(num, pval, has.zero, round = 3, option = r.flex.opts)
}
\arguments{
\item{num}{numeric - number}

\item{pval}{numeric - p value. If there is no need for stars this parameter should be set to NA}

\item{has.zero}{boolean - should it include leading zero (see APA 7 rules)}

\item{round}{integer - number of rounding places}

\item{option}{r.flex.opts}
}
\value{
string formated in a desired style
}
\description{
Converts number to APA 7 format and adds optional stars for pvalues
}
\examples{
# number higher than 1
star.pval.apa(num = 2.223, pval = 0.02, round = 3, has.zero = TRUE, option = list(d.p = "."))
# number lower than 1, no leading zero
star.pval.apa(num = 0.223, pval = 0.02, round = 3, has.zero = FALSE, option = list(d.p = "."))
# no p value
star.pval.apa(num = 0.223, pval = NA, round = 3, has.zero = FALSE, option = list(d.p = "."))
}
