% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{t.to.flex}
\alias{t.to.flex}
\title{Create flextable for t-test. A helper function for ttest.flex}
\usage{
\method{t}{to.flex}(rez, data, deci, effect, nvars, lang = "hr")
}
\arguments{
\item{rez}{a parameter}

\item{data}{a parameter}

\item{deci}{a parameter}

\item{effect}{a parameter}

\item{nvars}{a parameter}

\item{lang}{a parameter}
}
\value{
flextable
}
\description{
Create flextable for t-test. A helper function for ttest.flex
}
