% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_paired.R
\name{ttest.paired.flex}
\alias{ttest.paired.flex}
\title{Performs paired t-test for one or more pairs of variables.}
\usage{
ttest.paired.flex(data, ..., deci = 2, opts = r.flex.opts)
}
\arguments{
\item{data}{data frame with variables}

\item{...}{pairs of variables written as formulas and separated by comma. E.g. v1 ~ v2, v3 ~ v4}

\item{deci}{number of decimal places for M (default=2)}

\item{opts}{options - defaults to r.flex.opts}
}
\value{
list with
\itemize{
\item type - table type - used for inserting in word document
\item table - flextable with results
\item df - data frame with results as numerics
\item p.vals - pvalues for t-tests. This is used for future implementation of Benjamini-Hochbergs correction}
}
\description{
Performs paired t-test for one or more pairs of variables.
}
