% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_to_labelled.R
\name{var_to_labelled}
\alias{var_to_labelled}
\title{Add variable and value labels to a variable}
\usage{
var_to_labelled(data, vari, label = NULL, value_labs = NULL)
}
\arguments{
\item{data}{data frame}

\item{vari}{variable that needs to be changed as string}

\item{label}{label for the variable}

\item{value_labs}{named vector}
}
\value{
data frame with labelled variable
}
\description{
Add variable and value labels to a variable
}
\examples{
data.frame(a=c(1,1,1,2,2,2), b=c(1,2,3,1,2,3)) \%>\% var_to_labelled("a", "Age", c("1"="Young", "2"="Old"))
}
