% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{xtabs.flex}
\alias{xtabs.flex}
\title{Create crosstabs (contingency tables)}
\usage{
xtabs.flex(data, form, deci = 1, title = "", option = r.flex.opts)
}
\arguments{
\item{data}{dataset}

\item{form}{formula specified as row variables on left hand side and one section variable (columns) on right hand side. E.g. var1 + var2 + var3 ~ var_sect}

\item{deci}{number of decimal points (default=1)}

\item{title}{Table title (optional)}

\item{option}{inherits from r.flex.opts}
}
\value{
list with elements
\itemize{
\item type - table type - used for inserting in word document (, desc - are there descriptives)
\item title - used for table title. Can be set manually or automatically
\item section - section variable title
\item table - flextable with results
\item tab.df - results as data.frame
\item orientation - suggested page orientation (P/L)
}
}
\description{
Function creates crosstabulation with multiple variables in rows and one variable in column.
}
