% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_on_fire.R
\name{set.r.flex.opts}
\alias{set.r.flex.opts}
\title{Set r.flex.opts variable}
\usage{
set.r.flex.opts()
}
\value{
r.flex.opts environment used in various functions
}
\description{
Used as a direct way to set the options variable used in other functions. The preferred way is by using im_on_fire() function.
This should be used when you want to work without other perks of im_on_fire().
}
